% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiv_Score_Test_Region_multi.R
\name{Indiv_Score_Test_Region_multi}
\alias{Indiv_Score_Test_Region_multi}
\title{Multi-trait score test for individual variants in a given variant-set}
\usage{
Indiv_Score_Test_Region_multi(
  genotype,
  obj_nullmodel,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  rv_num_cutoff_max = 1e+09
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants.}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glmmkin_multi}} function for unrelated or
related samples. Note that \code{\link{fit_null_glmmkin_multi}}
is a wrapper of the \code{\link{glmmkin}} function from the \code{\link{GMMAT}} package.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{rv_num_cutoff_max}{the cutoff of maximum number of variants of analyzing
a given variant-set (default = 1e+09).}
}
\value{
a data frame with p rows corresponding to the p genetic variants in the given variant-set
and (k+1) columns: \code{Score1},...,\code{Scorek} (the score test statistics for each phenotype)
and \code{pvalue} (the multivariate score test p-value).
If a variant in the given variant-set has minor allele frequency = 0 or
greater than \code{rare_maf_cutoff}, the corresponding row will be \code{NA}. If a variant in
the given variant-set has degenerate covariance matrix across multiple phenotypes, the p-value will be set as 1.
}
\description{
The \code{Indiv_Score_Test_Region_multi} function takes in genotype and the object from fitting the null
model to analyze the associations between multiple (quantitative) phenotypes and
all individual variants in a given variant-set by using score test.
}
\references{
Li, X., Chen, H., et al. (2023). A statistical framework for powerful
multi-trait rare variant analysis in large-scale whole-genome sequencing studies.
\emph{bioRxiv}. (\href{https://doi.org/10.1101/2023.10.30.564764}{preprint})
}
